backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pandas
openpyxl
python-multipart